backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
python-dotenv
pandas
transformers
torch
google-generativeai
cachetools
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]